requirements.txt
asgiref==3.7.2
Django==3.2.20
imageio==2.31.1
networkx==2.6.3
numpy==1.21.6
packaging==23.1
Pillow==9.5.0
pytz==2023.3
PyWavelets==1.3.0
scikit-image==0.19.3
scipy==1.7.3
sklearn==0.0.post7
sqlparse==0.4.4
tifffile==2021.11.2
typing_extensions==4.7.1
